Refactor ECDSA structs into ec::signature
