+++ aws-lc-rs/src/signature.rs
@@ -255,8 +255,8 @@ use crate::rsa::{
 };
 
 pub use crate::ec::key_pair::{EcdsaKeyPair, PrivateKey as EcdsaPrivateKey};
-use crate::ec::EcdsaSignatureFormat;
-pub use crate::ec::{
+use crate::ec::signature::EcdsaSignatureFormat;
+pub use crate::ec::signature::{
     EcdsaSigningAlgorithm, EcdsaVerificationAlgorithm, PublicKey as EcdsaPublicKey,
 };
 pub use crate::ed25519::{
@@ -531,168 +531,168 @@ pub static RSA_PKCS1_SHA512: RsaSignatureEncoding = RsaSignatureEncoding::new(
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-256 curve and SHA-256.
 pub static ECDSA_P256_SHA256_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P256,
+    id: &ec::signature::AlgorithmID::ECDSA_P256,
     digest: &digest::SHA256,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-384 curve and SHA-384.
 pub static ECDSA_P384_SHA384_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P384,
+    id: &ec::signature::AlgorithmID::ECDSA_P384,
     digest: &digest::SHA384,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-384 curve and SHA3-384.
 pub static ECDSA_P384_SHA3_384_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P384,
+    id: &ec::signature::AlgorithmID::ECDSA_P384,
     digest: &digest::SHA3_384,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-521 curve and SHA-1.
 pub static ECDSA_P521_SHA1_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA1_FOR_LEGACY_USE_ONLY,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-521 curve and SHA-224.
 pub static ECDSA_P521_SHA224_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA224,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-521 curve and SHA-256.
 pub static ECDSA_P521_SHA256_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA256,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-521 curve and SHA-384.
 pub static ECDSA_P521_SHA384_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA384,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-521 curve and SHA-512.
 pub static ECDSA_P521_SHA512_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA512,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-521 curve and SHA3-512.
 pub static ECDSA_P521_SHA3_512_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA3_512,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-256K1 curve and SHA-256.
 pub static ECDSA_P256K1_SHA256_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P256K1,
+    id: &ec::signature::AlgorithmID::ECDSA_P256K1,
     digest: &digest::SHA256,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of fixed-length (PKCS#11 style) ECDSA signatures using the P-256K1 curve and SHA3-256.
 pub static ECDSA_P256K1_SHA3_256_FIXED: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P256K1,
+    id: &ec::signature::AlgorithmID::ECDSA_P256K1,
     digest: &digest::SHA3_256,
     sig_format: EcdsaSignatureFormat::Fixed,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-256 curve and SHA-256.
 pub static ECDSA_P256_SHA256_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P256,
+    id: &ec::signature::AlgorithmID::ECDSA_P256,
     digest: &digest::SHA256,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// *Not recommended.* Verification of ASN.1 DER-encoded ECDSA signatures using the P-256 curve and SHA-384.
 pub static ECDSA_P256_SHA384_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P256,
+    id: &ec::signature::AlgorithmID::ECDSA_P256,
     digest: &digest::SHA384,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// *Not recommended.* Verification of ASN.1 DER-encoded ECDSA signatures using the P-384 curve and SHA-256.
 pub static ECDSA_P384_SHA256_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P384,
+    id: &ec::signature::AlgorithmID::ECDSA_P384,
     digest: &digest::SHA256,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-384 curve and SHA-384.
 pub static ECDSA_P384_SHA384_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P384,
+    id: &ec::signature::AlgorithmID::ECDSA_P384,
     digest: &digest::SHA384,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-384 curve and SHA3-384.
 pub static ECDSA_P384_SHA3_384_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P384,
+    id: &ec::signature::AlgorithmID::ECDSA_P384,
     digest: &digest::SHA3_384,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-521 curve and SHA-1.
 pub static ECDSA_P521_SHA1_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA1_FOR_LEGACY_USE_ONLY,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-521 curve and SHA-224.
 pub static ECDSA_P521_SHA224_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA224,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-521 curve and SHA-256.
 pub static ECDSA_P521_SHA256_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA256,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-521 curve and SHA-384.
 pub static ECDSA_P521_SHA384_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA384,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-521 curve and SHA-512.
 pub static ECDSA_P521_SHA512_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA512,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-521 curve and SHA3-512.
 pub static ECDSA_P521_SHA3_512_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P521,
+    id: &ec::signature::AlgorithmID::ECDSA_P521,
     digest: &digest::SHA3_512,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-256K1 curve and SHA-256.
 pub static ECDSA_P256K1_SHA256_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P256K1,
+    id: &ec::signature::AlgorithmID::ECDSA_P256K1,
     digest: &digest::SHA256,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
 
 /// Verification of ASN.1 DER-encoded ECDSA signatures using the P-256K1 curve and SHA3-256.
 pub static ECDSA_P256K1_SHA3_256_ASN1: EcdsaVerificationAlgorithm = EcdsaVerificationAlgorithm {
-    id: &ec::AlgorithmID::ECDSA_P256K1,
+    id: &ec::signature::AlgorithmID::ECDSA_P256K1,
     digest: &digest::SHA3_256,
     sig_format: EcdsaSignatureFormat::ASN1,
 };
